Update GVDB
In particular, this adds gvdb_table_write_contents_async() from gvdb!4.

It also requires changing some integer types due to an API break in
gvdb_table_get_names().

+++ lib/contrib/gvdb/gvdb-reader.h
@@ -4,7 +4,7 @@
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
- * version 2 of the licence, or (at your option) any later version.
+ * version 2.1 of the License, or (at your option) any later version.
  *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -12,9 +12,7 @@
  * Lesser General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the
- * Free Software Foundation, Inc., 59 Temple Place - Suite 330,
- * Boston, MA 02111-1307, USA.
+ * License along with this library; if not, see <http://www.gnu.org/licenses/>.
  *
  * Author: Ryan Lortie <desrt@desrt.ca>
  */
@@ -31,23 +29,32 @@ G_BEGIN_DECLS
 GvdbTable *             gvdb_table_new_from_bytes                       (GBytes       *bytes,
                                                                          gboolean      trusted,
                                                                          GError      **error);
+
 GvdbTable *             gvdb_table_new                                  (const gchar  *filename,
                                                                          gboolean      trusted,
                                                                          GError      **error);
+
 void                    gvdb_table_free                                 (GvdbTable    *table);
+
 gchar **                gvdb_table_get_names                            (GvdbTable    *table,
-                                                                         gint         *length);
+                                                                         gsize        *length);
+
 gchar **                gvdb_table_list                                 (GvdbTable    *table,
                                                                          const gchar  *key);
+
 GvdbTable *             gvdb_table_get_table                            (GvdbTable    *table,
                                                                          const gchar  *key);
+
 GVariant *              gvdb_table_get_raw_value                        (GvdbTable    *table,
                                                                          const gchar  *key);
+
 GVariant *              gvdb_table_get_value                            (GvdbTable    *table,
                                                                          const gchar  *key);
 
+
 gboolean                gvdb_table_has_value                            (GvdbTable    *table,
                                                                          const gchar  *key);
+
 gboolean                gvdb_table_is_valid                             (GvdbTable    *table);
 
 G_END_DECLS
